Add distro specific variables
This is in preparation for using more distro specific information
since the existing repo is rather Debian/Ubuntu specific and
has some issues on RHEL/CentOS/Fedora as a result.

Would need to test on other distros and add their information
as well, but I think this covers most of the use cases now.

Signed-off-by: John 'Warthog9' Hawley <jhawley@vmware.com>
Signed-off-by: John 'Warthog9' Hawley <warthog9@eaglescrag.net>